Move migration code to a separate class
